+++ pkg-zid-proxy/files/usr/local/www/zid-proxy_groups.php
@@ -234,6 +234,7 @@ if ($csrf_token === '' && function_exists('csrf_get_token')) {
 // Display tabs
 $tab_array = array();
 $tab_array[] = array(gettext("Settings"), false, "/zid-proxy_settings.php");
+$tab_array[] = array(gettext("Active IPs"), false, "/zid-proxy_active_ips.php");
 $tab_array[] = array(gettext("Groups"), true, "/zid-proxy_groups.php");
 $tab_array[] = array(gettext("Access Rules"), false, "/zid-proxy_rules.php");
 $tab_array[] = array(gettext("Logs"), false, "/zid-proxy_log.php");